first separate between test and browser
